fix(auth): enable token invalidation and fix user ID type casting
- Update `SqliteUserRepo.UpdateUserAsync` to persist `tokens_valid_after` to the database, ensuring that token revocation (e.g., on logout) is permanent and enforced.

- Fix type casting in `AuthEndpoints` (`/auth/logout`) to correctly cast `UserId` as `long?` instead of `int?`, preventing potential null reference issues during logout.

+++ Backend/app/API/Endpoints/AuthEndpoints.cs
@@ -77,7 +77,7 @@ public static class AuthEndpoints
                 httpContext.Response.Cookies.Delete("auth_token", new CookieOptions { Path = "/" });
 
                 // If user is authenticated, invalidate all their tokens
-                var userId = httpContext.Items["UserId"] as int?;
+                var userId = httpContext.Items["UserId"] as long?;
                 if (userId.HasValue)
                 {
                     await authService.InvalidateAllTokensAsync(userId.Value);
